# src/swagger.yaml
openapi: 3.0.0
info:
  title: NoteMindAI App API
  version: "1.0.0"
  description: API for an AI-powered note-taking application.
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /summaries:
    get:
      summary: Retrieve all summaries
      tags: [Summaries]
      responses:
        '200':
          description: A list of summaries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Summary'
        '500':
          description: Internal server error
    post:
      summary: Create a new summary
      tags: [Summaries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Summary'
      responses:
        '200':
          description: Summary created successfully.
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /summaries/{noteId}:
    get:
      summary: Find a single summary by noteId
      tags: [Summaries]
      parameters:
        - in: path
          name: noteId
          required: true
          schema:
            type: integer
          description: ID of the note to retrieve its summary
      responses:
        '200':
          description: Summary details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '404':
          description: Summary not found
        '500':
          description: Internal server error
    put:
      summary: Update an existing summary by noteId
      tags: [Summaries]
      parameters:
        - in: path
          name: noteId
          required: true
          schema:
            type: integer
          description: ID of the note to update its summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Summary'
      responses:
        '200':
          description: Summary updated successfully.
        '404':
          description: Summary not found
        '500':
          description: Internal server error
    delete:
      summary: Remove a summary by noteId
      tags: [Summaries]
      parameters:
        - in: path
          name: noteId
          required: true
          schema:
            type: integer
          description: ID of the note to delete its summary
      responses:
        '200':
          description: Summary deleted successfully.
        '404':
          description: Summary not found
        '500':
          description: Internal server error
...
  /transcriptions:
    get:
      summary: Retrieve all transcriptions
      tags: [Transcriptions]
      responses:
        '200':
          description: A list of transcriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transcription'
        '500':
          description: Internal server error
    post:
      summary: Create a new transcription
      tags: [Transcriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transcription'
      responses:
        '200':
          description: Transcription created successfully.
        '400':
          description: Bad request
        '500':
          description: Internal server error
  /transcriptions/{id}:
    get:
      summary: Find a single transcription by ID
      tags: [Transcriptions]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the transcription to retrieve
      responses:
        '200':
          description: Transcription details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '404':
          description: Transcription not found
        '500':
          description: Internal server error
    put:
      summary: Update an existing transcription by ID
      tags: [Transcriptions]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the transcription to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transcription'
      responses:
        '200':
          description: Transcription updated successfully.
        '404':
          description: Transcription not found
        '500':
          description: Internal server error
    delete:
      summary: Remove a transcription by ID
      tags: [Transcriptions]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID of the transcription to delete
      responses:
        '200':
          description: Transcription deleted successfully.
        '404':
          description: Transcription not found
        '500':
          description: Internal server error
components:
  schemas:
...
    Summary:
      type: object
      properties:
        noteId:
          type: integer
          description: The ID of the note associated with this summary
        summary:
          type: string
          description: The text of the summary
      required:
        - noteId
        - summary
...
    Transcription:
      type: object
      properties:
        id:
          type: integer
          description: The ID of the transcription
        audioId:
          type: integer
          description: The ID of the related audio file
        text:
          type: string
          description: The transcribed text
        transcriptionQuality:
          type: number
          format: float
          description: The quality score of the transcription
      required:
        - audioId
        - text